+++ gdsfactory/functions.py
@@ -195,7 +195,7 @@ def add_settings_label(
     d = (
         {setting: component.get_setting(setting) for setting in settings}
         if settings
-        else component.info.changed
+        else component.metadata.changed
     )
 
     component.add_label(text=OmegaConf.to_yaml(d), layer=layer_label)
@@ -225,9 +225,12 @@ if __name__ == "__main__":
     )
     # c.show()
 
-    cr = c.rotate()
-    cr.pprint()
+    cr = rotate(component=c)
     cr.show()
+
+    # cr = c.rotate()
+    # cr.pprint()
+    # cr.show()
 
     # cm = move(c, destination=(20, 20))
     # cm.show()