+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # [CHANGELOG](https://keepachangelog.com/en/1.0.0/)
 
+## [5.6.4](https://github.com/gdsfactory/gdsfactory/pull/376)
+
+- add_fiber_array adds vertical ports to grating couplers
+- add_fiber_single adds vertical ports to grating couplers. Before it was adding only loopback ports.
+- import gds fixes [PR](https://github.com/gdsfactory/gdsfactory/pull/374)
+
 ## [5.6.3](https://github.com/gdsfactory/gdsfactory/pull/373)
 
 - fix get_labels rotation